requirement.txt
Django==3.1.6
django-crispy-forms==1.12.0
Jinja2==2.11.2
Pillow==8.1.0
virtualenv==20.4.6
pandas==1.3.2
scikit-learn==0.24.2
networkx==2.6.2
matplotlib==3.4.3
plotly==5.3.1
django-ckeditor==6.1.0
